+++ src/kaskas/io/providers/digital.hpp
@@ -16,7 +16,7 @@ public:
     DigitalSensor(const std::function<LogicalState()>& value_f) : _value_f(value_f){};
 
     LogicalState state() const { return _value_f(); }
-    double value() const { return _value_f(); }
+    float value() const { return _value_f(); }
 
     std::unique_ptr<prompt::RPCRecipe> rpc_recipe(const std::string_view& recipe_name,
                                                   const std::string_view& root) override {
@@ -47,7 +47,7 @@ public:
 
     LogicalState state() const { return _map.state_f(); }
     void set_state(LogicalState state) { _map.set_state_f(state); }
-    double value() const { return state(); }
+    float value() const { return state(); }
 
     std::unique_ptr<prompt::RPCRecipe> rpc_recipe(const std::string_view& recipe_name, const std::string_view& root) {
         return {};
